RooFit::MultiProcess & TestStatistics part 1: mathcore & Minuit2 changes (#8369)
* add virtuals to IFunction to calculate second derivative and stepsize

Used in Minuit2 ExternalInternalGradientCalculator (introduced in a later commit) and in RooFit (also later commits) to enable calculation of all derivative and related numbers outside of Minuit2. Specifically, having second derivative and step size calculated outside of Minuit2 allows exact replication of the gradients as calculated in Minuit2 itself, which in turn enables external optimization and parallelization of this process without having to modify Minuit2.

* add second derivative and step size calculation to Minuit2 function adaptors

Like the changes to IFunction in the previous commit, this is also necessary for Minuit2-external gradient calculation.

* add second derivative and step size to Minuit2 parameter transformers and increase their precision

This commit changes two (functionally speaking strictly separate, but unified in their ultimate purpose) things. First, to further accomodate the added ability of external gradient calculators to pass in second derivative and step sizes, the internal to external parameter transformers in Minuit2 need to have added D2Int2Ext and GStepInt2Ext functions that perform these transformations for those added parts. Second, we increase the precision of the transformers to long double. This is necessary  to make sure that trigonometric transformations don't lose a few bits of precision on back-and-forth parameter transformations, which they do when using the double versions. Both changes are necessary for the external gradient calculation in Minuit2 which will be introduced in a next commit in ExternalInternalGradientCalculator.

* implement optional second derivative and step size calculation in AnalyticalGradientCalculator

* remove unnecessary temporary objects

* add ExternalInternalGradientCalculator to minuit2

Derived class from AnalyticalGradientCalculator. Allows Minuit2 using
codes to calculate gradients themselves, but in Minuit2 "internal"
parameter space, in order to preserve exact value correspondence with
the Minuit2 implementation.

* Update math/mathcore/inc/Math/IFunction.h


* revert unnecessary math/minuit2 changes

As highlighted by @lmoneta in PR review comments (https://github.com/root-project/root/pull/8369#pullrequestreview-684824744), the second derivative and step size information is not necessary at this point. The floating point exact minuit-external (roofit-internal) derivative duplication is currently done using ExternalInternalGradientCalculator and further made possible by the long double parameter transformation functions. This is all we need for now.

* fix previous revert commit + revert one additional thing

Fixes: some merge mistakes + some remaining layout diffs.

The additional reverted thing is the block from line 119 in MnSeedGenerator.cxx.

* fix additional layout

Just trying to keep the PR diff minimal.

* remove copyright line [skip ci]

* add ExternalInternalGradientCalculator doxygen

+++ math/minuit2/inc/Minuit2/SqrtUpParameterTransformation.h
@@ -1,5 +1,5 @@
 // @(#)root/minuit2:$Id$
-// Authors: M. Winkler, F. James, L. Moneta, A. Zsenei   2003-2005
+// Authors: M. Winkler, F. James, L. Moneta, A. Zsenei, E.G.P. Bos   2003-2017
 
 /**********************************************************************
  *                                                                    *
@@ -36,13 +36,13 @@ public:
    ~SqrtUpParameterTransformation() {}
 
    // transformation from internal to external
-   double Int2ext(double Value, double Upper) const;
+   long double Int2ext(long double Value, long double Upper) const;
 
    // transformation from external to internal
-   double Ext2int(double Value, double Upper, const MnMachinePrecision &) const;
+   long double Ext2int(long double Value, long double Upper, const MnMachinePrecision &) const;
 
    // derivative of transformation from internal to external
-   double DInt2Ext(double Value, double Upper) const;
+   long double DInt2Ext(long double Value, long double Upper) const;
 
 private:
 };
