+++ math/minuit2/src/SinParameterTransformation.cxx
@@ -1,5 +1,5 @@
 // @(#)root/minuit2:$Id$
-// Authors: M. Winkler, F. James, L. Moneta, A. Zsenei   2003-2005
+// Authors: M. Winkler, F. James, L. Moneta, A. Zsenei, E.G.P. Bos   2003-2017
 
 /**********************************************************************
  *                                                                    *
@@ -16,25 +16,25 @@ namespace ROOT {
 
 namespace Minuit2 {
 
-double SinParameterTransformation::Int2ext(double Value, double Upper, double Lower) const
+long double SinParameterTransformation::Int2ext(long double Value, long double Upper, long double Lower) const
 {
    // transformation from  to internal (unlimited) to external values (limited by Lower/Upper )
    return Lower + 0.5 * (Upper - Lower) * (std::sin(Value) + 1.);
 }
 
-double
-SinParameterTransformation::Ext2int(double Value, double Upper, double Lower, const MnMachinePrecision &prec) const
+long double
+SinParameterTransformation::Ext2int(long double Value, long double Upper, long double Lower, const MnMachinePrecision &prec) const
 {
    // transformation from external (limited by Lower/Upper )  to internal (unlimited) values given the lower/upper
    // limits
 
-   double piby2 = 2. * std::atan(1.);
-   double distnn = 8. * std::sqrt(prec.Eps2());
-   double vlimhi = piby2 - distnn;
-   double vlimlo = -piby2 + distnn;
+   long double piby2 = 2. * std::atan(1.);
+   long double distnn = 8. * std::sqrt(prec.Eps2());
+   long double vlimhi = piby2 - distnn;
+   long double vlimlo = -piby2 + distnn;
 
-   double yy = 2. * (Value - Lower) / (Upper - Lower) - 1.;
-   double yy2 = yy * yy;
+   long double yy = 2. * (Value - Lower) / (Upper - Lower) - 1.;
+   long double yy2 = yy * yy;
    if (yy2 > (1. - prec.Eps2())) {
       if (yy < 0.) {
          // Lower limit
@@ -45,13 +45,12 @@ SinParameterTransformation::Ext2int(double Value, double Upper, double Lower, co
          //       std::cout<<"SinParameterTransformation warning: is at its Upper allowed limit."<<std::endl;
          return vlimhi;
       }
-
    } else {
       return std::asin(yy);
    }
 }
 
-double SinParameterTransformation::DInt2Ext(double Value, double Upper, double Lower) const
+long double SinParameterTransformation::DInt2Ext(long double Value, long double Upper, long double Lower) const
 {
    // return the derivative of the transformation d Ext/ d Int
    return 0.5 * ((Upper - Lower) * std::cos(Value));
